[DSC-651] Disable drag & drop on pending changes

+++ src/app/edit-item-relationships/edit-item-relationships.component.ts
@@ -121,6 +121,11 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
   isInit = false;
 
   /**
+   * Emits true when a relationship is being added, deleted, or updated
+   */
+  pendingChanges$ = new BehaviorSubject<boolean>(false);
+
+  /**
    * This parameter define the status of sidebar (hide/show)
    */
   private sidebarStatus$: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(true);
@@ -186,7 +191,8 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
    * Get all relationships of the relation to manage
    */
   retrieveRelationships(objectItem?: Item): void {
-    console.log('retrieveRelationships');
+    // TODO check
+    this.pendingChanges$.next(true);
     // this.subs.push(
       this.itemRD$.pipe(
         getRemoteDataPayload(),
@@ -207,6 +213,9 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
 
         this.updateStatusByItemId$.next(itemId);
         this.isInit = true;
+
+        // TODO check
+        this.pendingChanges$.next(false);
       });
       // );
   }
@@ -256,9 +265,15 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
         this.deleteAddRelationship(relationshipType, event.item, event.relationship);
       }
 
-    } else {
-      this.deleteRelationship(event.relationship, event.item);
+      return;
     }
+
+    if (event.action === 'delete' || event.action === 'unselect') {
+      this.deleteRelationship(event.relationship, event.item);
+      return;
+    }
+
+    console.warn(`Unhandled action ${event.action}`);
   }
 
   /**
@@ -267,9 +282,11 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
    * @param objectItem  the relationship type of the relationship created
    */
   addRelationship(type: RelationshipType, objectItem: Item): void {
+    this.pendingChanges$.next(true);
     this.relationshipService.addRelationship(type.id, objectItem, this.item, type.leftwardType, type.rightwardType).pipe(take(1))
       .subscribe(() => {
         this.retrieveRelationships(objectItem);
+        this.pendingChanges$.next(false);
       });
   }
 
@@ -278,8 +295,10 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
    * @param relationship  the relationship to update place
    */
   updateRelationship(relationship: Relationship): void {
+    this.pendingChanges$.next(true);
     this.relationshipService.updateRightPlace(relationship).pipe(take(1)).subscribe({
-      error: err => this.retrieveRelationships()
+      error: err => this.retrieveRelationships(), // TODO handle pending changes
+      complete: () => this.pendingChanges$.next(false)
     });
   }
 
@@ -289,8 +308,10 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
    * @param objectItem the reference to the objectItem that owns the relationship
    */
   deleteRelationship(relationship: Relationship, objectItem: Item): void {
+    this.pendingChanges$.next(true);
     this.relationshipService.deleteRelationship(relationship.id).subscribe((res) => {
       this.retrieveRelationships(objectItem);
+      this.pendingChanges$.next(false);
     });
   }
 
@@ -301,9 +322,11 @@ export class EditItemRelationshipsComponent implements OnInit, OnDestroy {
    * @param relationship  the relationship to delete
    */
   deleteAddRelationship(type: RelationshipType, objectItem: Item, relationship: Relationship) {
+    this.pendingChanges$.next(true);
     this.relationshipService.deleteRelationship(relationship.id).pipe(take(1))
       .subscribe(() => {
         this.addRelationship(type, objectItem);
+        this.pendingChanges$.next(false);
       });
   }
